refactor: ensure that estimator's test contract is always build
Replaces build.sh which has to be invoked manually with build.rs which
is run automatically when needed.
