+++ .buildkite/pipeline.yml
@@ -91,10 +91,6 @@ steps:
   - label: "runtime params estimate"
     command: |
       source ~/.cargo/env && set -eux
-      cd runtime/runtime-params-estimator/test-contract
-      rustup target add wasm32-unknown-unknown
-      ./build.sh
-      cd ..
       RUSTFLAGS='-D warnings' cargo run --release --package runtime-params-estimator --bin runtime-params-estimator -- --accounts-num 20000 --additional-accounts-num 200000 --iters 1 --warmup-iters 1 --metric time
 
     branches: "!master !beta !stable"
