+++ runtime/runtime-params-estimator/emu-cost/ci.sh
@@ -14,7 +14,6 @@ docker run \
      /usr/bin/env bash -c "
 set -ex
 cd /host/${srcdir}/runtime/runtime-params-estimator
-pushd test-contract; ./build.sh; popd
 cargo run --release --package neard --bin neard -- --home /tmp/data init --test-seed=alice.near --account-id=test.near --fast
 cargo run --release --package genesis-populate --bin genesis-populate -- --additional-accounts-num=200000 --home /tmp/data
 cargo build --release --package runtime-params-estimator --features required
