--- runtime/runtime-params-estimator/small-test-contract/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "smallest-contract"
-version = "0.1.0"
-authors = ["Near Inc <hello@nearprotocol.com>"]
-edition = "2018"
-
-[lib]
-crate-type = ["cdylib"]
-
-[profile.release]
-codegen-units = 1
-# Tell `rustc` to optimize for small code size.
-opt-level = "z"
-lto = true
-debug = false
-panic = "abort"
-
-[workspace]
-members = []
