+++ runtime/runtime-params-estimator/test-contract/.cargo/config
@@ -1,2 +1,2 @@
 [build]
-rustflags = ["-C", "link-args=-s"]
+rustflags = ["-C", "link-arg=-s"]
